split pip into two pips
